Add frame map
Maps a concrete stack frame to an artificial id that is passed to VS Code
as `frameId`.

Because the debugger protocol doesn't have any notion of frame ids we
calculate a unique identifier for each pair of context id and frame
depth and remember that in the map throughout the session. This way we
can give out frameIds to VS Code in a stackFrameRequest.