Wait after calling runScriptOnServer
Wait with a timeout after SDS connection is established and
runScriptOnServer is called before try connecting to the debugger port.

This commit fixes launch request for me.

The issue: Connecting to the debugger port regularly resulted in a
ECONNREFUSED error because the debugger engine was not ready on the
server. The script was executed and paused correctly but no TCP
listening socket had been set up so that the launch request failed.

Code desperately needs refactoring. Maybe retry a few time when
ECONNREFUSED is retrieved.