Add IPC to communicate back to extension
Use a UNIX Domain or Windows socket, respectively, to have a
communication channel from the debug adapter process back to the
extension.  This communication channel is outside of VS Code's debug
protocol and allows us to show input boxes or use other parts of the VS
Code extension API that would otherwise be unavailable to us in the
debug adapter.

Socket file path is usually /tmp/vscode-janus-debug.sock or so on Linux
and macOS.