Make sure Server Console always uses up-to-date connection parameters
Before this commit we were using
`vscode.workspace.getConfiguration('launch')` to load the launch
configurations. This function returned the currently active workspace
configuration which mustn't be necessarily the same as stored on disk.
In combination with the file watcher this resulted in old connection
parameters being used to connect to the remote server.