Add mapping from breakpoints of local files to breakpoints of remote files.
The JANUS server has a non standard feature that allows you to paste sources
together, but that mutates the line numbers and also the correct location of
the breakpoints.

This commit includes an algorithm to find out the displacement caused by the
pasting feature mentioned above to easy adjust the breakpoints.