Add ServerSource class
Maps the expanded, pre-processed sources as they exist in a JS context
to individual source files in the user's workspace. You'd use it as
this:

    let serverSource = ServerSource.fromSources(sourceLines);
    serverSource.toLocalPosition(12); // → { source: 'foo.js', line: 4 }

This allows us to calculate the original file and line positions in a
file that contains #import directives which got pre-processed by a
DOCUMENTS server.